Tests cleanup.
