--- handler/core/src/test/resources/task/customTaskProvider.conf
@@ -1,8 +1,0 @@
-# task factory
-factory = custom
-config {
-  anyKey = anyValue
-}
-# node & transitions
-graph {
-}
