--- handler/core/src/test/resources/task/defaultTaskProvider.conf
@@ -1,3 +1,0 @@
-# node & transitions
-graph {
-}
