--- handler/core/src/test/resources/task/defaultTaskProviderWithConfig.conf
@@ -1,6 +1,0 @@
-# node & transitions
-config {
-  testKey = testValue
-}
-graph {
-}
